+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Air Bullet spine: same as Bullet but floating in the air
 - Pink balancer: ``--visualize`` argument
 - PPO balancer: distribute sample policy